requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
redis==5.0.1
celery==5.3.4
openai==1.3.7
elevenlabs==0.2.26
deepgram==2.12.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
python-multipart==0.0.6
websockets==12.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
